[IMP] messages logic from bank class to its own class
